refactor(terminal): extract pane focus decoration into PaneDecorator
Moves the focus border + glow CALayer plumbing out of TerminalContainerView
into a standalone PaneDecorator helper. State enum now carries the workspace
tint as an associated value so the decorator stays oblivious to host state.
