+++ Bellith/Views/TerminalContainerView.swift
@@ -639,118 +639,20 @@ final class TerminalContainerView: NSView, TerminalOverlayControllerHost, Termin
                 } else if isBroadcasting {
                     state = .broadcast
                 } else if surface === activeSurface {
-                    state = .active
+                    state = .active(tint: activeWorkspaceTint())
                 } else {
                     state = .inactive
                 }
-                applyPaneDecoration(to: leaf, state: state)
+                PaneDecorator.apply(to: leaf, state: state)
             }
         }
 
         refreshPaneHeaders()
     }
 
-    private enum PaneDecorationState {
-        case hidden
-        case inactive
-        case active
-        case broadcast
-    }
-
-    private enum PaneDecoration {
-        static let border = "paneBorder"
-        static let glow = "paneGlow"
-        // Border sits flush with the pane card (the PaneContainerView paints
-        // its own card hairline at the same edge). The hairline is replaced
-        // by the active border's color here when the pane is focused.
-        static let inset: CGFloat = 0
-        static let cornerRadius: CGFloat = 8
-    }
-
-    private func applyPaneDecoration(to leaf: NSView, state: PaneDecorationState) {
-        leaf.wantsLayer = true
-        leaf.layer?.borderColor = nil
-        leaf.layer?.borderWidth = 0
-        leaf.layer?.cornerRadius = 0
-        leaf.layer?.masksToBounds = false
-
-        guard let layer = leaf.layer else { return }
-        let frame = leaf.bounds.insetBy(dx: PaneDecoration.inset, dy: PaneDecoration.inset)
-        let borderLayer = paneDecorationLayer(
-            named: PaneDecoration.border,
-            on: layer,
-            frame: frame
-        )
-        let glowLayer = paneDecorationLayer(
-            named: PaneDecoration.glow,
-            on: layer,
-            frame: frame
-        )
-
-        borderLayer.cornerRadius = PaneDecoration.cornerRadius
-        borderLayer.cornerCurve = .continuous
-        borderLayer.backgroundColor = NSColor.clear.cgColor
-        borderLayer.masksToBounds = false
-        glowLayer.cornerRadius = PaneDecoration.cornerRadius
-        glowLayer.cornerCurve = .continuous
-        glowLayer.backgroundColor = NSColor.clear.cgColor
-        glowLayer.borderWidth = 0
-
-        switch state {
-        case .hidden:
-            borderLayer.opacity = 0
-            glowLayer.opacity = 0
-            glowLayer.shadowOpacity = 0
-
-        case .inactive:
-            borderLayer.opacity = 1
-            borderLayer.borderWidth = 1
-            borderLayer.borderColor = Theme.chromeHairline.withAlphaComponent(Theme.colors.isLight ? 0.32 : 0.24).cgColor
-            glowLayer.opacity = 0
-            glowLayer.shadowOpacity = 0
-
-        case .active:
-            let tint = activeWorkspaceTint()
-            borderLayer.opacity = 1
-            borderLayer.borderWidth = 1.5
-            borderLayer.borderColor = tint.withAlphaComponent(0.34).cgColor
-            glowLayer.opacity = 1
-            glowLayer.shadowColor = tint.withAlphaComponent(0.12).cgColor
-            glowLayer.shadowOpacity = 0.7
-            glowLayer.shadowRadius = 10
-            glowLayer.shadowOffset = .zero
-
-        case .broadcast:
-            // Broadcast keeps the global accent — it's a destructive/global mode
-            // that shouldn't be tinted by workspace identity.
-            borderLayer.opacity = 1
-            borderLayer.borderWidth = 1.5
-            borderLayer.borderColor = Theme.accent.withAlphaComponent(0.52).cgColor
-            glowLayer.opacity = 1
-            glowLayer.shadowColor = Theme.accent.withAlphaComponent(0.12).cgColor
-            glowLayer.shadowOpacity = 1
-            glowLayer.shadowRadius = 8
-            glowLayer.shadowOffset = .zero
-        }
-    }
-
     private func activeWorkspaceTint() -> NSColor {
         guard selectedTabIndex < tabs.count else { return Theme.accent }
         return WorkspaceTint.accent(for: tabs[selectedTabIndex].title)
-    }
-
-    private func paneDecorationLayer(named name: String, on parent: CALayer, frame: CGRect) -> CALayer {
-        if let existing = parent.sublayers?.first(where: { $0.name == name }) {
-            existing.frame = frame
-            return existing
-        }
-
-        let layer = CALayer()
-        layer.name = name
-        layer.frame = frame
-        layer.autoresizingMask = [.layerWidthSizable, .layerHeightSizable]
-        parent.addSublayer(layer)
-        return layer
     }
 
     // MARK: - Tab Mode
